Added support for Coulomb Approx
Coulomb approximation has some problems related to the fact that it allows point charges to occupy the same space, giving rise to infinities.  This is a problem with the approximation scheme, not my handling of it.  I handled this by basically setting the infinities to zero and setting a flag to alert the user of a problem.

I did not integrate the Coulomb approximation into the main function-calling script, since I think it is of limited utility in solving problems.